+++ pallets/loans/src/tests/mod.rs
@@ -1,6 +1,8 @@
 use std::time::Duration;
 
 use cfg_mocks::pallet_mock_data::util::MockDataCollection;
+use cfg_primitives::SECONDS_PER_DAY;
+use cfg_traits::interest::{CompoundingSchedule, InterestRate};
 use cfg_types::permissions::{PermissionScope, PoolRole, Role};
 use frame_support::{assert_noop, assert_ok, storage::bounded_vec::BoundedVec};
 use sp_runtime::{traits::BadOrigin, DispatchError, FixedPointNumber};
