+++ pallets/loans/src/tests/mutate_loan.rs
@@ -1,5 +1,3 @@
-use cfg_traits::{CompoundingSchedule, InterestRate};
-
 use super::*;
 
 const DEFAULT_MUTATION: LoanMutation<Rate> = LoanMutation::InterestPayments(InterestPayments::None);
